test: migrate Phase 4 selection/gsheet non-Celery slices to component
Move auth/authz/validation/render/redirect coverage for sortition,
db-selection (legacy + backoffice) and gsheet selection routes to
tests/component over a FakeUnitOfWork. Celery dispatch/status-poll/revoke
tests stay in e2e; SelectionSettings round-trip + pagination stay
db_semantics. Add a component-tier Celery AsyncResult boundary stub to
conftest (the result backend is an external boundary the tier lacks),
keeping progress/status routes driven by the seeded SelectionRunRecord.
