+++ .secrets.baseline
@@ -567,5 +567,5 @@
       }
     ]
   },
-  "generated_at": "2026-06-23T09:32:55Z"
+  "generated_at": "2026-06-23T09:55:27Z"
 }
